requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
boto3==1.34.34
twilio==8.11.1
python-multipart==0.0.6
